add: plot.py
