+++ srcs/train.py
@@ -15,47 +15,7 @@ from srcs.init import he_normal, xavier_normal, normal
 from srcs.optimizer import SGD, Momentum, Nesterov, AdaGrad, RMSProp, Adam
 from srcs.layer import Dense
 from srcs.model import Sequential
-
-
-class RealtimePlot:
-    def __init__(self, max_iter):
-        self.max_iter = max_iter
-        self.iterations = []
-        self.train_losses = []
-        self.train_accs = []
-        self.valid_losses = []
-        self.valid_accs = []
-
-        self.fig, (self.ax1, self.ax2) = plt.subplots(2, 1, figsize=(10, 10))
-        plt.ion()  # インタラクティブモードをオン
-
-    def update(self, i, train_loss, train_acc, valid_loss, valid_acc):
-        self.iterations.append(i)
-        self.train_losses.append(train_loss)
-        self.train_accs.append(train_acc)
-        self.valid_losses.append(valid_loss)
-        self.valid_accs.append(valid_acc)
-
-        self._plot()
-        plt.pause(0.00001)
-
-    def _plot(self):
-        self.ax1.clear()
-        self.ax1.plot(self.iterations, self.train_losses, color='C0', linestyle='-', label='Train Loss')
-        self.ax1.plot(self.iterations, self.valid_losses, color='C1', linestyle='--', label='Valid Loss')
-        self.ax1.set_xlabel('Iterations')
-        self.ax1.set_ylabel('Loss')
-        self.ax1.legend()
-        self.ax1.set_xlim(0, self.max_iter)
-
-        self.ax2.clear()
-        self.ax2.plot(self.iterations, self.train_accs, color='C0', linestyle='-', label='Train Acc')
-        self.ax2.plot(self.iterations, self.valid_accs, color='C1', linestyle='--', label='Valid Acc')
-        self.ax2.set_xlabel('Iterations')
-        self.ax2.set_ylabel('Accuracy')
-        self.ax2.legend()
-        self.ax2.set_xlim(0, self.max_iter)
-        self.ax2.set_ylim(0, 1)
+from srcs.plot import RealtimePlot
 
 
 def train_model(
@@ -127,8 +87,7 @@ def train_model(
                 plotter.update(epoch, train_loss, train_acc, valid_loss, valid_acc)
 
     if plot:
-        plt.ioff()
-        plt.show()
+        plotter.plot()
 
     return iterations, train_losses, train_accs, valid_losses, valid_accs
